Enhance custom portrait functionality and update documentation
- Updated the custom portrait fetching script to support full emotion sets for starter evolutions and specific Pokémon, improving visual representation in-game.
- Modified the runtime configuration to clarify the custom portrait feature, ensuring accurate descriptions in the codebase.
- Enhanced documentation to reflect the new coverage of custom portraits, including details on supported Pokémon and emotion slots.
- Removed outdated portrait files for various Pokémon, streamlining asset management and reducing clutter in the graphics directory.

+++ include/custom_portraits.h
@@ -6,7 +6,7 @@
 extern const FileArchive gCustomPortraitArchive;
 extern const s16 gCustomPortraitSpecies[];
 extern const u16 gCustomPortraitMasks[];
-extern const u8 gCustomPortraitCount;
+extern const u16 gCustomPortraitCount;
 
 /* TRUE if species has a SpriteCollab portrait pack in gCustomPortraitArchive. */
 bool8 HasCustomPortrait(s16 speciesId);
